requirements.txt
Django==3.0.7
django-crispy-forms==1.9.1
django-extensions==2.2.9
django-seed==0.2.2
djangorestframework==3.11.0
